remove vestiges of gnomemenu, prettify __init__.py
Removes gnomemenu from __init__.py and namcap.1

In __init__.py, each module is on a separate line in
alphabetical order.  This makes removal and addition of
modules easier to spot in the diffs.  I've used split() here
so that I can do without the quotes and commas in the list.

Signed-off-by: Abhishek Dasgupta <abhidg@gmail.com>
Signed-off-by: Hugo Doria <hugo@archlinux.org>