Make the depends module not suck
Namcap being slow pissed me off enough that I wanted to figure out what
could be taking so long. As is usual with software, 1 module was causing
about 90% of the slowdown. Compare the before and after with this patch:

$ time namcap -m -r depends {4 packages} >/dev/null
real	0m18.001s
user	0m9.313s
sys	0m7.966s

$ time ./namcap.py -m -r depends {4 packages }>/dev/null
real	0m4.512s
user	0m3.770s
sys	0m0.740s

This patch addresses a few issue noticed when crawling and profiling the code
using cProfile. First, we were doing O(n*m) string compares of files in the
local database with libraries in the package. We can use a few tricks to cut
down the number of checks- namely, once a library has been found we don't need
to look for it anymore, and we can also ignore any non-'.so' files.

Another part where we were really shooting ourselves in the foot was invoking
readelf on *every single file* in the package. Cut that down so we only call
readelf on files with the ELF magic bytes, which gives us a dramatic savings as
we no longer fork for files in /usr/share, for instance.

Bonus material: script checking left out perl, and script dependencies were
formerly listed as link-level dependencies as well. Both of these have been
fixed.

The sad part is there is still room to improve here, but I decided enough
was enough for now. :)

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Hugo Doria <hugo@archlinux.org>