Use the is_elf function in other modules where it makes sense
The rpath module is another one that exhibits rather poor behavior, as it
also tried to call readelf on things that didn't make sense at all. Squash
this and get some big gains on packages with a lot of files.

As a point of reference, running all modules on 4 varying packages has gone
from 29.7 seconds to 5.0 seconds.

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Hugo Doria <hugo@archlinux.org>