setup.py: Make sdist command work
I have no idea how the last namcap release tarball was built, but sdist was
busted because it didn't include in the tarball a lot of the files we try to
install. Fix it up by adding a MANIFEST.in file used by distutils.

Now, to make a tarball, one can just do:
  ./setup.py sdist

Signed-off-by: Dan McGee <dan@archlinux.org>