Add new 'anyelf' rule
I can't think of too many valid situations where an ELF object in an 'any'
package makes sense, so make a rule that calls this out as a namcap error.
This rule comes about due to FS#18747 and a python library binding package
being invalidly converted to the 'any' architecture.

Signed-off-by: Dan McGee <dan@archlinux.org>