Handle optdepends description stripping correctly
Commit 95986ec3 attempted to fix handling of optdepends with descriptions,
but it only affected the PKGBUILD parser. Do the description stripping in
the package building code in namcap proper so we can handle all varieties of
packages, whether from a PKGBUILD, the database, or a package file. The
original fix is reverted as part of this commit, and this fixes FS#18259.

Signed-off-by: Dan McGee <dan@archlinux.org>