Make is_elf() more resilient to weird perms
Some files in packages do not have the owner read bit set, which means
open() blows up when trying to peek into the file. Add a workaround that
temporarily sets the extracted file to be readable so we can peek at it and
determine if it is an ELF file. For those files that are either links or
something else funky, just assume it is not an ELF file.

Signed-off-by: Dan McGee <dan@archlinux.org>