parsepkgbuild: don't let test [ return codes leak
when using `[ ] && action` style tests, the return code of the test can
leak, which was unfortuante for every PKGBUILD that didn't contain an
install file as the last test would always return 1 (failure). We want
parsepkgbuild to return 0 if we make it all the way through.

"Fix" all of the shortcut test forms to just use the full `if [ ]; then`
block instead of a shortcut.

Signed-off-by: Dan McGee <dan@archlinux.org>