Revert "Reduce the amount of false positives in extravars rule."
This reverts the non-test portions of commit
e2044d4583f80cd6283a2f8d0ffa1d6f5d834123.

This is not the proper way to fix this, and we most definitely should
not have to parse, subprocess, or do anything with bash more than once,
especially in a rule. We also shouldn't assume everyone has a perfectly
clean environment with no extra variables exported and defined. Revert
the entire thing, minus the added test, with a better fix on the way.

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Rémy Oudompheng <remy@archlinux.org>