Move the extravars work to the PKGBUILD parser
Add an additional fake package variable '%SETVARS%' that shows the
currently defined variables when processing a PKGBUILD. Use this list in
the extravars rule instead of broke-ass regex parsing that will never
catch every possible corner case.

Also update the tests accordingly. Variables defined inside a function
are nothing we are going to worry about at the moment, so safely ignore
them. Add a legal variable to ensure we don't have any false positives,
and fix our own misgivings in the PKGBUILD parser so we aren't polluting
the namespace.

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Rémy Oudompheng <remy@archlinux.org>